Updated license info and general layout.
Replaced copy of LGPLv2.1 text with the official plaintext version
(https://www.gnu.org/licenses/old-licenses/lgpl-2.1.txt) and added in
the CC0 license in prep for the merge of newer PDCLib code (which is
explicitly CC0 instead of public domain).

Moved files to be more consistently named with the 'standard' layout
used in other projects.