Updated PDCLib and fixed licensing.
Updated PDCLib from most recent upstream, both because it hadn't been
updated in a _long_ time (close to 3-4 years, I believe), and to clarify
which files we have significantly modified and which are essentially
the same as upstream. Any files we have significantly modified are
relicensed (retaining an attribution note) as LGPLv2.1 to be consistent
with the library, but any that we have not retain their CC0 license.